+++ R/init_agents.R
@@ -2,10 +2,13 @@
 #' @description Creates and initializes agents with specified attributes, actions, and other parameters.
 #'
 #' @param n An integer specifying the number of agents to create. If \code{NULL}, \code{n} is derived from \code{attr_df} if provided.
-#' @param attr_df A data frame or a vector containing the initial attributes for each agent. Each row represents an agent's attributes.
-#' @param act_FUN A function or list of functions defining the actions or methods for each agent.
-#' @param active_binding_field A character vector specifying fields that should have active bindings.
+#' @param attr_df A data frame or a vector containing the initial attributes for each agent.
+#' Each row represents an agent's attributes.
+#' @param act_FUN A function or list of functions defining the actions or
+#' methods for each agent.
+#' @param active_binding A function or list of functions defining the output of the active bindings.
 #' @param ID_start An integer specifying the starting value for agent IDs. Default is \code{1}.
+#' @param custom_ID Optional vector of agent IDs. If provided, its length must match \code{n}.
 #' @param other_attrs A named list of additional attributes for each agent. Each list element must have the same length as \code{n}.
 #'
 #' @return A list of initialized agent objects, each with an assigned ID and specified attributes and actions.
@@ -14,23 +17,33 @@
 #' The \code{init_agents} function generates agent instances with specified attributes
 #' and functions.
 #'
-#' Number of agents:
+#' **Number of agents**:
 #' Either \code{n} or \code{attr_df} must be provided to specify the number of agents.
 #' If both \code{n} and \code{attr_df} are given, the function checks that
 #' the number of agents (\code{n}) matches the number of rows in \code{attr_df}.
 #'
-#' act_FUN: When a single function is provided, it is replicated across all agents,
+#' **act_FUN**:
+#' When a single function is provided, it is replicated across all agents,
 #' meaning that each agent follows the same behavior rule.
 #' If you want each agent to have a different behavior, pass a vector or list of
 #' functions, each corresponding to an agent. For example, to assign three different
 #' functions \code{act_x1}, \code{act_x2}, and \code{act_x3} to three agents respectively,
-#' set \code{act_FUN = list(act1 = list(act_x1, act_x2, act_x3))}.
+#' set:
+#'
+#' \code{act_FUN = list(act1 = list(act_x1, act_x2, act_x3))}
+#'
 #' To assign multiple actions to each agent, note that the first level of
 #' the list will be shared across all agents.
 #' For example, to assign two actions, \code{act_x} and \code{act_y}, to all agents,
-#' set \code{act_FUN = list(act_x = act_x, act_y = act_y)}.
+#' set:
+#'
+#' \code{act_FUN = list(act_x = act_x, act_y = act_y)}.
+#'
 #' To assign unique actions for each agent individually,
-#' use a structure like \code{act_FUN = list(act_x = list(act_x2, act_x2, act_x3), act_y = list(act_y1, act_y2, act_y3))}.
+#' use a structure like:
+#'
+#' \code{act_FUN = list(act_x = list(act_x2, act_x2, act_x3),
+#'                      act_y = list(act_y1, act_y2, act_y3))}.
 #'
 #' You can specify each agent's behavior rule in four ways:
 #' - Function object: The basic approach, such as \code{act_FUN = act_x1}.
@@ -42,11 +55,9 @@
 #'  e.g., \code{act_FUN = "act_x1(a = 1)"}.
 #'
 #' Automatically, \code{G = G} and \code{E = E} are added to the function arguments
-#' within \code{act_FUN} to facilitate simulation in the \code{runABM} function.
-#' Additionally, \code{self <- self} is inserted at the beginning of the function
-#' body to enable agents to reference themselves within embedded functions.
+#' within \code{act_FUN} and \code{active_binding} to facilitate simulation in the \code{runABM} function.
 #'
-#' active_binding_field:
+#' **active_binding**:
 #' An active binding field recalculates its value based on other fields of
 #' the agent according to a predefined function.
 #' The specification follows the same format as for \code{act_FUN}.
@@ -54,20 +65,43 @@
 #' the desired field value. For more information,
 #' see the R6 package tutorial: <https://r6.r-lib.org/articles/Introduction.html>.
 #'
-#' ID_start:
+#' Example of active binding:
+#' Suppose you want each agent to have a `status` field that automatically returns
+#'  `"adult"` if the agent's `age >= 18`, and `"child"` otherwise.
+#'  You can define this behavior as follows:
+#'
+#' \code{
+#' active_binding <- list(
+#'   status = function() {
+#'     if (self$age >= 18) {
+#'       "adult"
+#'     } else {
+#'       "child"
+#'     }
+#'   }
+#' )
+#' }
+#'
+#' Note: In this example, `status` is not stored as a regular field;
+#' it is dynamically computed from the `age` field each time it is accessed.
+#'
+#' Important considerations for active binding:
+#' - Active bindings depend on other fields (e.g., `self$age` in the above example).
+#' These dependent fields must be defined before the active binding is accessed.
+#' Otherwise, an error such as "object 'age' not found" will occur.
+#' - You can reference other fields using `self$` within the function body.
+#' - Be cautious not to create circular dependencies (e.g., `status` depends on
+#' `score`, and `score` depends on `status`).
+#'
+#' **ID_start**:
 #' Each agent in a simulation should have a unique ID. If using multiple types of
 #' agents within a simulation,
 #' set this value to avoid ID overlap between agent groups.
-#' On the other hand, this means that the list index and ID_number does not match anymore,
-#' which is sometimes cumbersome for writing a selection function.
-#' Generally, it is recommended to leave the ID_start as the default, because
-#' during the simulation, it is specified which sets of agents to pickup and hence
-#' overlap of the IDs among the sets of the agents is not problematic.
 #'
-#' other_attrs:
+#' **other_attrs**:
 #' The primary way to specify agent attributes is through \code{attr_df}, a data frame.
-#' However, if an agent's attribute needs to be a vector, matrix,
-#' or data frame (e.g., for individual agent-level data), use \code{other_attrs}
+#' However, if an agent's attribute needs to be a vector, matrix, data frame, or list
+#'  (e.g., for individual agent-level data), use \code{other_attrs}
 #' as a list. Ensure each attribute within \code{other_attrs} has \code{n} elements
 #' to match the number of agents.
 #' Similar to \code{act_FUN}, the first list level represents fields.
@@ -85,49 +119,46 @@
 #'
 
 init_agents <- function(n = NULL, attr_df = NULL, act_FUN = NULL,
-                       active_binding_field = NULL, ID_start = 1,
+                       active_binding = NULL, ID_start = 1,
+                       custom_ID = NULL,
                        other_attrs = NULL){
-  # attrを処理する
+  # Process attr_df
   attr_sbs <- substitute(attr_df)
   attr_df <- .shape_agent_attr(attr = attr_df, attr_sbs = attr_sbs)
 
-  # nが投入されている場合には、数値かどうかを判定
+  # Check if n is an integer
   if(!is.null(n)){
-    stopifnot("n must be numeric." = is.numeric(n)==TRUE)
+    stopifnot("n must be numeric." = is.numeric(n))
   }
 
-  # nとattrの投入状況ごとに整理
+  # check the integration between n and attr_df
   if(!is.null(n) & !is.null(attr_df)){
-    ## nとattrが両方投入されている場合：両者が同じかをチェック
-    stopifnot("n must be the same to nrow in each attr." = n == NROW(attr_df))
+    ## When both n and attr_df are provided: check if they are the same
+    stopifnot("'n' must match the number of rows in 'attr_df'." = n == NROW(attr_df))
   }else if(is.null(n) & !is.null(attr_df)){
-    ## nはなし、attrは投入あり：nをattrから取得
+    ## n is NULL and attr_df presents: get n from attr_df
     n <- NROW(attr_df)
   }else if(!is.null(n) & is.null(attr_df)){
-    ## nはあり、attr_dfはなし：nはそのまま
+    ## n is present and attr_df is NULL: remain n as it is
     n <- n
   }else{
-    ## nもattrもなし：エラーを出す
-    stop("Either n or attr must be inputted.")
+    ## Neither n and attr_df exists: Express ERROR
+    stop("Either 'n' or 'attr_df' must be provided.")
   }
 
-  # act_FUNを整理
+  # Process act_FUN
   act_FUN_sbs <- substitute(act_FUN)
   act_FUN_list <- .shape_act_FUN(act_FUN = act_FUN,
                                  act_FUN_sbs = act_FUN_sbs,
                                  n = n)
 
-  # active_binding_fieldを整理
-  if(!is.null(active_binding_field)){
-    active_binding_field_sbs <- substitute(active_binding_field)
-    active_binding_field_formatted <- .shape_active_binding_field_agent(
-      active_binding_field = active_binding_field,
-      active_binding_field_sbs = active_binding_field_sbs)
-  }else{
-    active_binding_field_formatted <- NULL
-  }
+  # Process active_binding_field
+  active_binding_sbs <- substitute(active_binding)
+  active_binding_list <- .shape_act_FUN(act_FUN = active_binding,
+                                        act_FUN_sbs = active_binding_sbs,
+                                        n = n)
 
-  # G = G, E = E
+  # Add G = G, E = E to formals
   act_FUN_list <- lapply(act_FUN_list, function(FUN_vec){
     lapply(FUN_vec, function(FUN){
       # Delete G if the user already wrote it.
@@ -138,56 +169,78 @@ init_agents <- function(n = NULL, attr_df = NULL, act_FUN = NULL,
     })
   })
 
+  active_binding_list <- lapply(active_binding_list, function(FUN_vec){
+    lapply(FUN_vec, function(FUN){
+      # Delete G if the user already wrote it.
+      current_formals <- formals(FUN)
+      current_formals[which(names(current_formals)=="G")|which(names(current_formals)=="E")] <- NULL
+      formals(FUN) <- c(alist(G = G, E = E), current_formals) # add G=G, E = E
+      FUN
+    })
+  })
+
   # agentID
-  agentID <- ID_start:(ID_start + n - 1)
+  if(!is.null(custom_ID)){
+    stopifnot("The length of 'custom_ID' must match the number of agents." = length(custom_ID) == n)
+    agentID <- custom_ID
+  }else{
+    agentID <- seq(ID_start, length.out = n)
+  }
 
   # other_attrs-------
   if(!is.null(other_attrs)){
     other_attrs_sbs <- substitute(other_attrs)
-    ## インプットの形式の確認
+
+    ## check the input(data.frame is not allowed)
     stopifnot("other_attrs must be a list." = is.list(other_attrs) & !is.data.frame(other_attrs))
-    ## 中身の値の確認
+
+    ## check the contents
     lapply(other_attrs, function(X){
-      stopifnot("The length of each element within '...' must be n." = n == length(X))
+      stopifnot("Each element in 'other_attrs' must have length 'n'." = length(X) == n)
       if(is.function(X)){
-        stop("Elements passed through '...' must be cannot be functions.")
+        stop("Elements in 'other_attrs' must not be functions.")
       }
     })
-    ## ラベルの処理
-    other_attrs_character <- as.character(other_attrs_sbs)[-1]
-    other_attrs_obs_name <- sapply(other_attrs_character, function(X){
-      parsed_X <- parse(text = X)[[1]]
-      if(is.symbol(parsed_X)){
-        as.character(X)
+
+    ## get unnamed labels from substitutes
+    other_attrs_expr <- as.list(other_attrs_sbs)[-1] # drop the function name
+    other_attrs_obs_name <- sapply(other_attrs_expr, function(expr){
+      if(is.symbol(expr)){
+        deparse(expr)
       }else{
-        return(NA)
+        NA_character_
       }
     })
+
+    ## Integrate unnamed and named elements
     other_attrs_label <- names(other_attrs)
     if(is.null(other_attrs_label)){
       other_attrs_label <- other_attrs_obs_name
     }else{
       other_attrs_label[other_attrs_label==""] <- other_attrs_obs_name[other_attrs_label==""]
     }
-    other_attrs_label[is.na(other_attrs_label)] <- paste0("Z", 1:length(other_attrs_label[is.na(other_attrs_label)]))
+
+    # append the labels to unnamed objects
+    other_attrs_label[is.na(other_attrs_label)] <- paste0("Z", seq_len(sum(is.na(other_attrs_label))))
+
+    # append the labels
     names(other_attrs) <- other_attrs_label
   }
 
-  # attr_dfとdotをまとめる
+  # append attr_df and other_attrs
   attr_list <- c(as.list(attr_df), other_attrs)
 
-  # attr_listとact_FUN_listの名前が重複していないかをチェック
-  name_table <- table(c(names(attr_list),names(act_FUN_list)))
+  # check if attr_listとact_FUN_listの名前が重複していないかをチェック
+  name_table <- table(c(names(attr_list), names(act_FUN_list), names(active_binding_list)))
   duplicated_field_name <- names(name_table)[name_table>1]
   if(length(duplicated_field_name) > 0){
     stop(paste0("The field name '", duplicated_field_name, "' is duplicated. Please provide a unique name for each field."))
   }
 
-  # agentのR6クラスを作成
+  # create an R6 class for agents
   act_FUN_len <- length(act_FUN_list)
   act_FUN_i <- vector("list", act_FUN_len)
   names(act_FUN_i) <- names(act_FUN_list)
-
 
   out <- lapply(1:n, function(i){
     ID_i <- c(ID = agentID[i])
@@ -204,24 +257,24 @@ init_agents <- function(n = NULL, attr_df = NULL, act_FUN = NULL,
         act_FUN_i[[p]] <- act_FUN_list[[p]][[i]]
       }
     }
-    # 付加する
+    # add to Agents
     Agent <- ABM_Agent$new(fields = c(ID_i, attr_i), methods = act_FUN_i)
 
-    # active_binding_fieldを処理する
-    if(!is.null(active_binding_field_formatted)){
-      active <- assign_func_envs(active_binding_field_formatted, Agent$.__enclos_env__)
+    # process active_bindings
+    if(!is.null(active_binding_list)){
+      active_binding_i <- lapply(names(active_binding_list), function(X){active_binding_list[[X]][[i]]})
+      names(active_binding_i) <- names(active_binding_list)
+      active <- assign_func_envs(active_binding_i, Agent$.__enclos_env__)
       for(name in names(active)){
         makeActiveBinding(name, active[[name]], Agent$.__enclos_env__$self)
       }
       Agent$.__enclos_env__$.__active__ <- active
     }
-    # リターン
+    # return
     Agent
   })
 
-  # IDを名前として付与する
-  names(out) <- paste0("ID", agentID)
-  # リターン
+  # return
   out
 }